Address PR comments. Move parameters package to add.parameters; change copyright to 2016.
