+++ services/ps/bin/run.sh
@@ -1,5 +1,5 @@
 #!/bin/sh
-# Copyright (C) 2015 Seoul National University
+# Copyright (C) 2016 Seoul National University
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
